apps/recorder/pyproject.toml
[project]
name = "recorder"
version = "0.1.0"
description = "Standalone data recorder for Bybit mainnet capture"
requires-python = ">=3.11"
dependencies = [
    "event-saver",
    "bybit-adapter",
    "grid-db",
    "pyyaml>=6.0",
    "pydantic>=2.0",
]

[tool.uv.sources]
event-saver = { workspace = true }
bybit-adapter = { workspace = true }
grid-db = { workspace = true }

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/recorder"]
